WIP: split place controller tests (checkpoint)
